+++ lab04_painter/test/lib/unit/mock/MockCanvas.cpp
@@ -8,5 +8,5 @@ class MockCanvas : public ICanvas
 public:
 	MOCK_METHOD(void, SetColor, (Color color), (override));
 	MOCK_METHOD(void, Drawline, (const Point2D& from, const Point2D& to), (override));
-	MOCK_METHOD(void, DrawEllipse, (const Point2D& from, int horizontalRadius, int verticalRadius), (override));
+	MOCK_METHOD(void, DrawEllipse, (const Point2D& from, double horizontalRadius, double verticalRadius), (override));
 };
